refactor(frontends/basic): extract lowering pipeline helpers
